# static/api.yaml
openapi: '3.0.0'
info:
  title: Decentraland Governance API
  version: '1.0.0'
servers:
  - url: localhost:4000/api
...
paths:
...
  /updates/financials:
    get:
      tags:
        - Project Updates
      summary: Get all financial records
      parameters:
        - name: page_size
          in: query
          description: Page size
          schema:
            $ref: '#/components/schemas/limitOffset'
        - name: page_number
          in: query
          description: Page number
          schema:
            $ref: '#/components/schemas/limitOffset'
      responses:
        '200':
          $ref: '#/components/responses/200'
        '400':
          $ref: '#/components/responses/400'
  /updates/{update}:
    get:
      tags:
        - Project Updates
      summary: Get the update details from the given update ID
      parameters:
        - name: update
          $ref: '#/components/parameters/updateId'
      responses:
        '200':
          $ref: '#/components/responses/200'
  /updates/{update}/comments:
    get:
      tags:
        - Project Updates
      summary: Get the update comments from the given update ID
      parameters:
        - name: update
          $ref: '#/components/parameters/updateId'
      responses:
        '200':
          $ref: '#/components/responses/200'
...
  /votes:
    get:
      tags:
        - Votes
      summary: Get the cached votes from the given proposal ID
      parameters:
        - name: id
          in: query
          description: Proposal ID
          required: true
          schema:
            type: string
      responses:
        '200':
          $ref: '#/components/responses/200'
...
  /votes/{address}:
    get:
      tags:
        - Votes
      summary: Get the votes from a given address
      parameters:
        - name: address
          $ref: '#/components/parameters/address'
      responses:
        '200':
          $ref: '#/components/responses/200'
  /votes/top-voters:
    get:
      tags:
        - Votes
      summary: Get the top voters for the last 30 days
      parameters:
        - name: limit
          in: query
          required: false
          description: >-
            How many voters to return. Must be a whole number between 1 and 100;
            anything else is rejected with a 400. Defaults to 5 when omitted, but an
            empty value is not the same as omitting it and is rejected.
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 5
      responses:
        '400':
          description: The limit was not a whole number between 1 and 100
        '200':
          $ref: '#/components/responses/200'
...
  /proposals/{proposal}/votes:
    get:
      tags:
        - Votes
      summary: Get the votes from a given proposal ID
      parameters:
        - name: proposal
          $ref: '#/components/parameters/proposalId'
      responses:
        '200':
          $ref: '#/components/responses/200'
...
  /user/{address}/is-validated:
    get:
      tags:
        - Users
      summary: Check if the given address is validated
      parameters:
        - name: address
          $ref: '#/components/parameters/address'
        - name: account
          in: query
          required: true
          description: >-
            Which linked account to check. Repeat the parameter to ask about more
            than one, in which case every one of them must be linked for the
            answer to be true. Every value must be recognised: if any is not, the
            request is rejected rather than answered about the rest.
          schema:
            type: array
            minItems: 1
            items:
              type: string
              enum:
                - forum
                - discord
                - push
      responses:
        '200':
          $ref: '#/components/responses/200'
        '400':
          description: No account was given, or one of them was not a recognised account type
...
components:
...
  schemas:
...
    ethAddress:
      type: string
      pattern: '^0x[a-fA-F0-9]{40}$'
...
    limitOffset:
      type: integer
      minimum: 0
...
  parameters:
...
    address:
      name: address
      in: path
      description: User eth address
      required: true
      schema:
        $ref: '#/components/schemas/ethAddress'
    proposalId:
      name: proposal
      in: path
      description: Proposal ID
      required: true
      schema:
        type: string
    updateId:
      name: update
      in: path
      description: Update ID
      required: true
      schema:
        type: string
...
  responses:
    '200':
      description: successful operation
    '400':
      description: bad request
...